game-flow: merge M_HandleLevelComplete to trx
