+++ src/libtrx/game/game_flow/sequencer_events.c
@@ -17,6 +17,7 @@
 #include "version.h"
 
 static DECLARE_GF_EVENT_HANDLER(M_HandleExitToTitle);
+static DECLARE_GF_EVENT_HANDLER(M_HandleLevelComplete);
 static DECLARE_GF_EVENT_HANDLER(M_HandlePlayLevel);
 static DECLARE_GF_EVENT_HANDLER(M_HandlePlayCutscene);
 static DECLARE_GF_EVENT_HANDLER(M_HandlePlayFMV);
@@ -30,6 +31,7 @@ static DECLARE_GF_EVENT_HANDLER(M_HandleSetupBaconLara);
 static DECLARE_GF_EVENT_HANDLER((*m_EventHandlers[GFS_NUMBER_OF])) = {
     // clang-format off
     [GFS_EXIT_TO_TITLE]     = M_HandleExitToTitle,
+    [GFS_LEVEL_COMPLETE]    = M_HandleLevelComplete,
     [GFS_LOOP_GAME]         = M_HandlePlayLevel,
     [GFS_PLAY_CUTSCENE]     = M_HandlePlayCutscene,
     [GFS_PLAY_FMV]          = M_HandlePlayFMV,
@@ -54,6 +56,38 @@ static DECLARE_GF_EVENT_HANDLER((*m_EventHandlers[GFS_NUMBER_OF])) = {
 static DECLARE_GF_EVENT_HANDLER(M_HandleExitToTitle)
 {
     return (GF_COMMAND) { .action = GF_EXIT_TO_TITLE };
+}
+
+static DECLARE_GF_EVENT_HANDLER(M_HandleLevelComplete)
+{
+    if (seq_ctx != GFSC_NORMAL) {
+        return (GF_COMMAND) { .action = GF_NOOP };
+    }
+    const GF_LEVEL *const current_level = Game_GetCurrentLevel();
+    const GF_LEVEL *const next_level = GF_GetLevelAfter(current_level);
+
+    if (current_level == GF_GetLastLevel()) {
+        g_Config.profile.new_game_plus_unlock = true;
+        Config_Update();
+    }
+
+    RESUME_INFO *const resume = Savegame_GetCurrentInfo(current_level);
+    resume->flags.available = true;
+    const bool bonus_level_unlock = Stats_CheckAllSecretsCollected(GFL_NORMAL);
+
+    if (next_level != nullptr) {
+        Savegame_PersistGameToCurrentInfo(next_level);
+    }
+    if (next_level == nullptr) {
+        return (GF_COMMAND) { .action = GF_NOOP };
+    }
+    if (next_level->type == GFL_BONUS && !bonus_level_unlock) {
+        return (GF_COMMAND) { .action = GF_EXIT_TO_TITLE };
+    }
+    return (GF_COMMAND) {
+        .action = GF_START_GAME,
+        .param = next_level->num,
+    };
 }
 
 static DECLARE_GF_EVENT_HANDLER(M_HandlePlayLevel)
