+++ src/tr1/game/game_flow/sequencer_events.c
@@ -11,42 +11,13 @@
 #include <libtrx/game/option/passport.h>
 #include <libtrx/game/rooms.h>
 
-static DECLARE_GF_EVENT_HANDLER(M_HandleLevelComplete);
 static DECLARE_GF_EVENT_HANDLER(M_HandleDisableFloor);
 
 static DECLARE_GF_EVENT_HANDLER((*m_EventHandlers[GFS_NUMBER_OF])) = {
     // clang-format off
-    [GFS_LEVEL_COMPLETE]   = M_HandleLevelComplete,
     [GFS_DISABLE_FLOOR]    = M_HandleDisableFloor,
     // clang-format on
 };
-
-static DECLARE_GF_EVENT_HANDLER(M_HandleLevelComplete)
-{
-    if (seq_ctx != GFSC_NORMAL) {
-        return (GF_COMMAND) { .action = GF_NOOP };
-    }
-    const GF_LEVEL *const current_level = Game_GetCurrentLevel();
-    const GF_LEVEL *const next_level = GF_GetLevelAfter(current_level);
-
-    if (current_level == GF_GetLastLevel()) {
-        g_Config.profile.new_game_plus_unlock = true;
-        Config_Update();
-    }
-    const bool bonus_level_unlock = Stats_CheckAllSecretsCollected(GFL_NORMAL);
-
-    if (next_level == nullptr) {
-        return (GF_COMMAND) { .action = GF_NOOP };
-    }
-
-    if (next_level->type == GFL_BONUS && !bonus_level_unlock) {
-        return (GF_COMMAND) { .action = GF_EXIT_TO_TITLE };
-    }
-    return (GF_COMMAND) {
-        .action = GF_START_GAME,
-        .param = next_level->num,
-    };
-}
 
 static DECLARE_GF_EVENT_HANDLER(M_HandleDisableFloor)
 {
