+++ src/tr2/game/stats.h
@@ -1,6 +1,5 @@
 #pragma once
 
-#include <libtrx/game/game_flow.h>
 #include <libtrx/game/items/types.h>
 #include <libtrx/game/stats.h>
 
@@ -9,6 +8,5 @@ FINAL_STATS Stats_ComputeFinalStats(GF_LEVEL_TYPE level_type);
 
 int32_t Stats_GetMaxSecrets(void);
 bool Stats_CheckAllLevelSecretsCollected(void);
-bool Stats_CheckAllSecretsCollected(GF_LEVEL_TYPE level_type);
 uint32_t Stats_ReserveSecretBit(OBJECT_ID object_id);
 OBJECT_ID Stats_GetSecretObject(int32_t secret_idx);
